Restructure to avoid import issues
