+++ Procfile
@@ -1,2 +1,2 @@
 release: flask db upgrade
-web: gunicorn src.app:create_app\(\) -b 0.0.0.0:$PORT -w 3
+web: gunicorn app:create_app\(\) -b 0.0.0.0:$PORT -w 3
