+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "CHIME-2",
   "version": "1.0.0",
-  "description": "Covid-19 Hospital Impact Model for Epidemics",
+  "description": "Covid-19 Hospital Impact Model for Epidemics 2",
   "scripts": {
     "build": "run-script-os",
     "build:win32": "SET NODE_ENV=production && webpack --progress --colors -p && npm run flask-static-digest",
