+++ test/test_forms.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """Test forms."""
 
-from src.public.forms import LoginForm
-from src.user.forms import RegisterForm
+from app.public.forms import LoginForm
+from app.user.forms import RegisterForm
 
 
 class TestRegisterForm:
