+++ test/test_functional.py
@@ -5,7 +5,7 @@ See: http://webtest.readthedocs.org/
 """
 from flask import url_for
 
-from src.user.models import User
+from app.user.models import User
 
 from .factories import UserFactory
